TeensyLC friendly, remove fast_log.h, use AUDIO_SAMPLE_RATE / 10 instead of 4410
